feat: teamlist 페이지 추가 및 팀 페이지 구조 개선 및 SEO 최적화
