+++ app/components/Team/Member.tsx
