+++ app/layout.tsx
@@ -1,9 +1,39 @@
 import "./globals.css";
 import type { Metadata } from "next";
 
+// metadataBase는 상대 경로를 절대 URL로 변환하기 위한 기본 URL 설정
 export const metadata: Metadata = {
-  title: "Coworkers",
-  description: "팀 협업 관리 서비스",
+  metadataBase: new URL("https://coworkers.com"),
+  title: {
+    default: "Coworkers - 팀 협업 관리 서비스",
+    template: "%s | Coworkers", // 하위 페이지에서 %s가 title로 대체됨
+  },
+  description:
+    "팀 협업을 위한 할 일 관리, 멤버 초대, 프로젝트 진행 상황 추적 서비스",
+  keywords: ["팀 협업", "프로젝트 관리", "할 일 관리", "업무 관리", "팀워크"],
+  authors: [{ name: "Coworkers" }],
+  creator: "Coworkers",
+  openGraph: {
+    type: "website",
+    locale: "ko_KR",
+    url: "/",
+    title: "Coworkers - 팀 협업 관리 서비스",
+    description:
+      "팀 협업을 위한 할 일 관리, 멤버 초대, 프로젝트 진행 상황 추적 서비스",
+    siteName: "Coworkers",
+    images: [
+      {
+        url: "/opengraph-image.png",
+        width: 1200,
+        height: 630,
+        alt: "Coworkers",
+      },
+    ],
+  },
+  robots: {
+    index: true,
+    follow: true,
+  },
 };
 
 export default function RootLayout({
